server/requirements.txt
arabic-reshaper==2.1.3
asgiref==3.6.0
auditlog3==1.0.1
autopep8==1.5.4
Babel==2.12.1
black==23.3.0
certifi==2021.5.30
charset-normalizer==3.1.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.2.1
django-ckeditor==6.1.0
django-colorfield==0.4.1
django-crispy-forms==1.10.0
django-easy-pdf==0.1.1
django-import-export==2.5.0
django-js-asset==2.0.0
django-model-utils==4.1.1
django-notifications-hq==1.6.0
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-softdelete==0.9.2
django-utils-six==2.0
djangorestframework==3.14.0
enum34==1.1.10
et-xmlfile==1.0.1
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
jsonfield==3.1.0
MarkupPy==1.14
MarkupSafe==2.1.2
mypy-extensions==1.0.0
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.7
packaging==23.1
pathspec==0.11.1
Pillow==8.2.0
platformdirs==3.5.1
pycodestyle==2.6.0
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.4.1
reportbro-fpdf==1.7.10
reportbro-lib==1.6.0
reportlab==3.5.67
requests==2.31.0
selenium==3.141.0
sentry-sdk==1.1.0
simpleeval==0.9.13
simplejson==3.19.1
six==1.15.0
sqlparse==0.4.1
swapper==1.1.2.post1
tablib==3.0.0
toml==0.10.2
tomli==2.0.1
typing==3.7.4.3
uritemplate==4.1.1
urllib3==1.26.6
webencodings==0.5.1
xhtml2pdf==0.2.5
xlrd==2.0.1
XlsxWriter==1.3.9
xlwt==1.3.0
